Create separate ScrapeRateMyProfessors class
